Connection and Handler Paradigm2
