--- meerk40t/device/ch341/ch341handler.py
@@ -1,9 +1,0 @@
-
-
-class CH341Handler:
-    def __init__(self, channel, state):
-        self.channel = channel
-        self.state = state
-
-    def connect(self, index=0):
-        pass
