+++ meerk40t/device/ch341/mock.py
@@ -3,8 +3,8 @@
 
 import time
 
-from meerk40t.device.ch341.ch341connection import CH341Connection
-from meerk40t.device.ch341.ch341handler import CH341Handler
+from .ch341 import Connection as CH341Connection
+from .ch341 import Handler as CH341Handler
 
 
 class CH341Driver(CH341Connection):
