+++ meerk40t/device/moshi/moshicontroller.py
@@ -40,6 +40,8 @@ class MoshiController(Module):
         self._usb_state = -1
 
         self.ch341 = self.context.open("module/ch341")
+        self.connection = None
+
         self.max_attempts = 5
         self.refuse_counts = 0
         self.connection_errors = 0
@@ -79,7 +81,7 @@ class MoshiController(Module):
 
         @self.context.console_command("usb_disconnect", help="Disconnect USB")
         def usb_disconnect(command, channel, _, args=tuple(), **kwargs):
-            if self.ch341.driver is not None:
+            if self.connection is not None:
                 self.close()
             else:
                 channel("Usb is not connected.")
@@ -176,27 +178,28 @@ class MoshiController(Module):
         self.realtime_pipe(swizzle_table[1][0])
 
     def realtime_pipe(self, data):
-        self.ch341.write_addr(data)
+        self.connection.write_addr(data)
 
     def open(self):
         self.pipe_channel("open()")
-        if self.ch341.driver is None:
-            self.ch341.detect_driver_and_open()
+        if self.connection is None:
+            self.connection = self.ch341.connect(self.context.mock)
         else:
             # Update criteria
-            self.ch341.index = self.context.usb_index
-            self.ch341.bus = self.context.usb_bus
-            self.ch341.address = self.context.usb_address
-            self.ch341.serial = self.context.usb_serial
-            self.ch341.chipv = self.context.usb_version
-            self.ch341.open()
-        if self.ch341 is None:
+            # TODO: RESTORE CRITERIA
+            self.connection.index = self.context.usb_index
+            self.connection.bus = self.context.usb_bus
+            self.connection.address = self.context.usb_address
+            self.connection.serial = self.context.usb_serial
+            self.connection.chipv = self.context.usb_version
+            self.connection.open()
+        if self.connection is None:
             raise ConnectionRefusedError
 
     def close(self):
         self.pipe_channel("close()")
-        if self.ch341.driver is not None:
-            self.ch341.close()
+        if self.connection is not None:
+            self.connection.close()
 
     def control(self, command):
         if command == "execute\n":
@@ -408,11 +411,11 @@ class MoshiController(Module):
         return True  # A packet was prepped and sent correctly.
 
     def send_packet(self, packet):
-        self.ch341.write(packet)
+        self.connection.write(packet)
         self.update_packet(packet)
 
     def update_status(self):
-        self._status = self.ch341.get_status()
+        self._status = self.connection.get_status()
         if self.context is not None:
             self.context.signal(
                 "pipe;status",
